+++ app/src/timepicker/utils/index.ts
@@ -166,3 +166,17 @@ export const timeConversion = (str = '') => {
 
   return `${hours}:${mins}`;
 };
+
+export const debounce = <T extends (...args: any[]) => ReturnType<T>>(
+  callback: T,
+  timeout: number,
+): ((...args: Parameters<T>) => void) => {
+  let timer: ReturnType<typeof setTimeout>;
+
+  return (...args: Parameters<T>) => {
+    clearTimeout(timer);
+    timer = setTimeout(() => {
+      callback(...args);
+    }, timeout);
+  };
+};