+++ app/src/timepicker/utils/variables.ts
@@ -1,8 +1,5 @@
-const name = 'timepicker-ui';
-const mouseEvents = 'mousedown mouseup mousemove mouseleave mouseover';
-const touchEvents = 'touchstart touchmove touchend';
-const allEvents: string = mouseEvents.concat(` ${touchEvents}`);
-const selectorActive = 'active';
-
-// eslint-disable-next-line object-curly-newline
-export { name, mouseEvents, touchEvents, allEvents, selectorActive };
+export const name = 'timepicker-ui';
+export const mouseEvents = 'mousedown mouseup mousemove mouseleave mouseover';
+export const touchEvents = 'touchstart touchmove touchend';
+export const allEvents = `${mouseEvents} ${touchEvents}`;
+export const selectorActive = 'active';